Split ECDSA and RSA functions into their own methods and add a base level 'XXXGenerate' function so we can do more testing
