+++ cmd/gen-key/main.go
@@ -22,267 +22,23 @@
 package main
 
 import (
-	"crypto"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rsa"
-	"crypto/sha256"
 	"crypto/x509"
-	"encoding/asn1"
-	"encoding/binary"
 	"encoding/pem"
 	"errors"
 	"flag"
 	"fmt"
 	"log"
-	"math/big"
 	"os"
 
 	"github.com/miekg/pkcs11"
 )
 
 type Ctx interface {
+	GenerateKeyPair(pkcs11.SessionHandle, []*pkcs11.Mechanism, []*pkcs11.Attribute, []*pkcs11.Attribute) (pkcs11.ObjectHandle, pkcs11.ObjectHandle, error)
 	GetAttributeValue(pkcs11.SessionHandle, pkcs11.ObjectHandle, []*pkcs11.Attribute) ([]*pkcs11.Attribute, error)
 	SignInit(pkcs11.SessionHandle, []*pkcs11.Mechanism, pkcs11.ObjectHandle) error
 	Sign(pkcs11.SessionHandle, []byte) ([]byte, error)
 	GenerateRandom(pkcs11.SessionHandle, int) ([]byte, error)
-}
-
-func rsaArgs(label string, modLen, exp uint) ([]*pkcs11.Mechanism, []*pkcs11.Attribute, []*pkcs11.Attribute) {
-	expSlice := make([]byte, 8)
-	binary.BigEndian.PutUint64(expSlice, uint64(exp))
-	log.Printf("\tEncoded public exponent (%d) as: %0X\n", exp, expSlice)
-	return []*pkcs11.Mechanism{
-			pkcs11.NewMechanism(pkcs11.CKM_RSA_PKCS_KEY_PAIR_GEN, nil),
-		},
-		[]*pkcs11.Attribute{
-			pkcs11.NewAttribute(pkcs11.CKA_LABEL, label),
-			pkcs11.NewAttribute(pkcs11.CKA_TOKEN, true),
-			pkcs11.NewAttribute(pkcs11.CKA_VERIFY, true),
-			pkcs11.NewAttribute(pkcs11.CKA_MODULUS_BITS, modLen),
-			pkcs11.NewAttribute(pkcs11.CKA_PUBLIC_EXPONENT, expSlice),
-		}, []*pkcs11.Attribute{
-			pkcs11.NewAttribute(pkcs11.CKA_LABEL, label),
-			pkcs11.NewAttribute(pkcs11.CKA_TOKEN, true),
-			// Prevent attributes being retrieved
-			pkcs11.NewAttribute(pkcs11.CKA_SENSITIVE, true),
-			// Prevent the key being extracted from the device
-			pkcs11.NewAttribute(pkcs11.CKA_EXTRACTABLE, false),
-			// Allow the key to sign data
-			pkcs11.NewAttribute(pkcs11.CKA_SIGN, true),
-		}
-}
-
-func rsaPub(ctx Ctx, session pkcs11.SessionHandle, object pkcs11.ObjectHandle, modLen, exp uint) (*rsa.PublicKey, error) {
-	// Retrieve the public exponent and modulus for the generated public key
-	attrs, err := ctx.GetAttributeValue(session, object, []*pkcs11.Attribute{
-		pkcs11.NewAttribute(pkcs11.CKA_PUBLIC_EXPONENT, nil),
-		pkcs11.NewAttribute(pkcs11.CKA_MODULUS, nil),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("Failed to retrieve key attributes: %s", err)
-	}
-
-	// Attempt to build the public key from the retrieved attributes
-	pubKey := &rsa.PublicKey{}
-	gotMod, gotExp := false, false
-	for _, a := range attrs {
-		switch a.Type {
-		case pkcs11.CKA_PUBLIC_EXPONENT:
-			pubKey.E = int(big.NewInt(0).SetBytes(a.Value).Int64())
-			if pubKey.E != int(exp) {
-				return nil, errors.New("Returned CKA_PUBLIC_EXPONENT doesn't match expected exponent")
-			}
-			gotExp = true
-			log.Printf("\tPublic exponent: %d\n", pubKey.E)
-		case pkcs11.CKA_MODULUS:
-			pubKey.N = big.NewInt(0).SetBytes(a.Value)
-			if pubKey.N.BitLen() != int(modLen) {
-				fmt.Println("asd", pubKey.N.BitLen())
-				return nil, errors.New("Returned CKA_MODULUS isn't of the expected bit length")
-			}
-			gotMod = true
-			log.Printf("\tModulus: (%d bits) %X\n", pubKey.N.BitLen(), pubKey.N.Bytes())
-		}
-	}
-	// Fail if we are missing either the public exponent or modulus
-	if !gotExp || !gotMod {
-		return nil, errors.New("Couldn't retrieve modulus or exponent")
-	}
-	return pubKey, nil
-}
-
-func rsaVerify(ctx Ctx, session pkcs11.SessionHandle, object pkcs11.ObjectHandle, pub *rsa.PublicKey) error {
-	// Initialize a signing operation
-	err := ctx.SignInit(session, []*pkcs11.Mechanism{pkcs11.NewMechanism(pkcs11.CKM_RSA_PKCS, nil)}, object)
-	if err != nil {
-		return fmt.Errorf("Failed to initialize signing operation: %s", err)
-	}
-	input := []byte{0x30, 0x31, 0x30, 0x0d, 0x06, 0x09, 0x60, 0x86, 0x48, 0x01, 0x65, 0x03, 0x04, 0x02, 0x01, 0x05, 0x00, 0x04, 0x20} // SHA-256 prefix
-	msg, err := getRandomBytes(ctx, session)
-	if err != nil {
-		return fmt.Errorf("Failed to retrieve random data: %s", err)
-	}
-	log.Printf("\tConstructed random number: %d (%X)\n", big.NewInt(0).SetBytes(msg), msg)
-	hash := sha256.Sum256(msg)
-	log.Printf("\tMessage SHA-256 hash: %X\n", hash)
-	input = append(input, hash[:]...)
-	log.Println("\tSigning message")
-	signature, err := ctx.Sign(session, input)
-	if err != nil {
-		return fmt.Errorf("Failed to sign data: %s", err)
-	}
-	log.Printf("\tMessage signature: %X\n", signature)
-	err = rsa.VerifyPKCS1v15(pub, crypto.SHA256, hash[:], signature)
-	if err != nil {
-		return fmt.Errorf("Failed to verify signature: %s", err)
-	}
-	log.Println("\tSignature verified")
-	return nil
-}
-
-var stringToCurve = map[string]elliptic.Curve{
-	"P224": elliptic.P224(),
-	"P256": elliptic.P256(),
-	"P384": elliptic.P384(),
-	"P521": elliptic.P521(),
-}
-
-var curveToOID = map[elliptic.Curve]asn1.ObjectIdentifier{
-	elliptic.P224(): asn1.ObjectIdentifier{1, 3, 132, 0, 33},
-	elliptic.P256(): asn1.ObjectIdentifier{1, 2, 840, 10045, 3, 1, 7},
-	elliptic.P384(): asn1.ObjectIdentifier{1, 3, 132, 0, 34},
-	elliptic.P521(): asn1.ObjectIdentifier{1, 3, 132, 0, 35},
-}
-
-func ecArgs(label string, curve elliptic.Curve) ([]*pkcs11.Mechanism, []*pkcs11.Attribute, []*pkcs11.Attribute) {
-	encodedCurve, err := asn1.Marshal(curveToOID[curve])
-	if err != nil {
-		panic(err)
-	}
-	log.Printf("\tEncoded curve parameters as: %X\n", encodedCurve)
-	return []*pkcs11.Mechanism{
-			pkcs11.NewMechanism(pkcs11.CKM_EC_KEY_PAIR_GEN, nil),
-		}, []*pkcs11.Attribute{
-			pkcs11.NewAttribute(pkcs11.CKA_LABEL, label),
-			pkcs11.NewAttribute(pkcs11.CKA_TOKEN, true),
-			pkcs11.NewAttribute(pkcs11.CKA_VERIFY, true),
-			pkcs11.NewAttribute(pkcs11.CKA_EC_PARAMS, encodedCurve),
-		}, []*pkcs11.Attribute{
-			pkcs11.NewAttribute(pkcs11.CKA_LABEL, label),
-			pkcs11.NewAttribute(pkcs11.CKA_TOKEN, true),
-			// Prevent attributes being retrieved
-			pkcs11.NewAttribute(pkcs11.CKA_SENSITIVE, true),
-			// Prevent the key being extracted from the device
-			pkcs11.NewAttribute(pkcs11.CKA_EXTRACTABLE, false),
-			// Allow the key to sign data
-			pkcs11.NewAttribute(pkcs11.CKA_SIGN, true),
-		}
-}
-
-var oidDERToCurve = map[string]elliptic.Curve{
-	"06052B81040021":       elliptic.P224(),
-	"06082A8648CE3D030107": elliptic.P256(),
-	"06052B81040022":       elliptic.P384(),
-	"06052B81040023":       elliptic.P521(),
-}
-
-func ecPub(ctx Ctx, session pkcs11.SessionHandle, object pkcs11.ObjectHandle, curve elliptic.Curve) (*ecdsa.PublicKey, error) {
-	// Retrieve the curve and public point for the generated public key
-	attrs, err := ctx.GetAttributeValue(session, object, []*pkcs11.Attribute{
-		pkcs11.NewAttribute(pkcs11.CKA_EC_PARAMS, nil),
-		pkcs11.NewAttribute(pkcs11.CKA_EC_POINT, nil),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("Failed to retrieve key attributes: %s", err)
-	}
-
-	pubKey := &ecdsa.PublicKey{}
-	gotCurve, gotPoint := false, false
-	for _, a := range attrs {
-		switch a.Type {
-		case pkcs11.CKA_EC_PARAMS:
-			rCurve, present := oidDERToCurve[fmt.Sprintf("%X", a.Value)]
-			if !present {
-				return nil, errors.New("Unknown curve OID value returned")
-			}
-			pubKey.Curve = rCurve
-			if pubKey.Curve != curve {
-				return nil, errors.New("Returned CKA_EC_PARAMS doesn't match expected curve")
-			}
-			gotCurve = true
-		case pkcs11.CKA_EC_POINT:
-			x, y := elliptic.Unmarshal(curve, a.Value)
-			if x == nil {
-				// http://docs.oasis-open.org/pkcs11/pkcs11-curr/v2.40/os/pkcs11-curr-v2.40-os.html#_ftn1
-				// PKCS#11 v2.20 specified that the CKA_EC_POINT was to be stored in a DER-encoded
-				// OCTET STRING.
-				var point asn1.RawValue
-				_, err = asn1.Unmarshal(a.Value, &point)
-				if err != nil {
-					return nil, fmt.Errorf("Failed to unmarshal returned CKA_EC_POINT: %s", err)
-				}
-				if len(point.Bytes) == 0 {
-					return nil, errors.New("Invalid CKA_EC_POINT value returned")
-				}
-				x, y = elliptic.Unmarshal(curve, point.Bytes)
-				if x == nil {
-					fmt.Println(point.Bytes)
-					return nil, errors.New("Invalid CKA_EC_POINT value returned")
-				}
-			}
-			pubKey.X, pubKey.Y = x, y
-			gotPoint = true
-			log.Printf("\tX: %X\n", pubKey.X.Bytes())
-			log.Printf("\tY: %X\n", pubKey.Y.Bytes())
-		}
-	}
-	if !gotPoint || !gotCurve {
-		return nil, errors.New("Couldn't retrieve EC point or EC parameters")
-	}
-	return pubKey, nil
-}
-
-var curveToHash = map[elliptic.Curve]crypto.Hash{
-	elliptic.P224(): crypto.SHA256,
-	elliptic.P256(): crypto.SHA256,
-	elliptic.P384(): crypto.SHA384,
-	elliptic.P521(): crypto.SHA512,
-}
-
-var hashToString = map[crypto.Hash]string{
-	crypto.SHA256: "SHA-256",
-	crypto.SHA384: "SHA-384",
-	crypto.SHA512: "SHA-512",
-}
-
-func ecVerify(ctx Ctx, session pkcs11.SessionHandle, object pkcs11.ObjectHandle, pub *ecdsa.PublicKey) error {
-	err := ctx.SignInit(session, []*pkcs11.Mechanism{pkcs11.NewMechanism(pkcs11.CKM_ECDSA, nil)}, object)
-	if err != nil {
-		return fmt.Errorf("failed to initialize signing operation: %s", err)
-	}
-	msg, err := getRandomBytes(ctx, session)
-	if err != nil {
-		return fmt.Errorf("failed to retrieve random data: %s", err)
-	}
-	log.Printf("\tConstructed random number: %d (%X)\n", big.NewInt(0).SetBytes(msg), msg)
-	hashFunc := curveToHash[pub.Curve].New()
-	hashFunc.Write(msg)
-	hash := hashFunc.Sum(nil)
-	log.Printf("\tMessage %s hash: %X\n", hashToString[curveToHash[pub.Curve]], hash)
-	signature, err := ctx.Sign(session, hash)
-	if err != nil {
-		return fmt.Errorf("failed to sign data: %s", err)
-	}
-	log.Printf("\tMessage signature: %X\n", signature)
-	r := big.NewInt(0).SetBytes(signature[:len(signature)/2])
-	s := big.NewInt(0).SetBytes(signature[len(signature)/2:])
-	fmt.Println("post", r, s)
-	if !ecdsa.Verify(pub, hash[:], r, s) {
-		return errors.New("failed to verify ECDSA signature over test data")
-	}
-	log.Println("\tSignature verified")
-	return nil
 }
 
 func getRandomBytes(ctx Ctx, session pkcs11.SessionHandle) ([]byte, error) {
@@ -291,6 +47,29 @@ func getRandomBytes(ctx Ctx, session pkcs11.SessionHandle) ([]byte, error) {
 		return nil, err
 	}
 	return r, nil
+}
+
+func initialize(module string, slot uint, pin string) (Ctx, pkcs11.SessionHandle, error) {
+	ctx := pkcs11.New(module)
+	if ctx == nil {
+		return nil, 0, errors.New("failed to load module")
+	}
+	err := ctx.Initialize()
+	if err != nil {
+		return nil, 0, fmt.Errorf("couldn't initialize context: %s", err)
+	}
+
+	session, err := ctx.OpenSession(slot, pkcs11.CKF_SERIAL_SESSION|pkcs11.CKF_RW_SESSION)
+	if err != nil {
+		return nil, 0, fmt.Errorf("couldn't open session: %s", err)
+	}
+
+	err = ctx.Login(session, pkcs11.CKU_USER, pin)
+	if err != nil {
+		return nil, 0, fmt.Errorf("couldn't login: %s", err)
+	}
+
+	return ctx, session, nil
 }
 
 func main() {
@@ -320,23 +99,9 @@ func main() {
 		panic("--label is required")
 	}
 
-	ctx := pkcs11.New(*module)
-	if ctx == nil {
-		panic("failed to load module")
-	}
-	err := ctx.Initialize()
+	ctx, session, err := initialize(*module, *slot, *pin)
 	if err != nil {
-		panic(err)
-	}
-
-	session, err := ctx.OpenSession(*slot, pkcs11.CKF_SERIAL_SESSION|pkcs11.CKF_RW_SESSION)
-	if err != nil {
-		panic(err)
-	}
-
-	err = ctx.Login(session, pkcs11.CKU_USER, *pin)
-	if err != nil {
-		panic(err)
+		log.Fatalf("Failed to setup and session PKCS#11 context: %s", err)
 	}
 
 	var pubKey interface{}
@@ -345,54 +110,18 @@ func main() {
 		if *rsaModLen == 0 {
 			panic("--modulus-bits is required")
 		}
-		log.Printf("Generating RSA key with %d bit modulus and public exponent %d\n", *rsaModLen, *rsaExp)
-		m, pubTmpl, privTmpl := rsaArgs(*label, *rsaModLen, *rsaExp)
-		pub, priv, err := ctx.GenerateKeyPair(session, m, pubTmpl, privTmpl)
+		pubKey, err = rsaGenerate(ctx, session, *label, *rsaModLen, *rsaExp)
 		if err != nil {
-			panic(err)
+			log.Fatalf("Failed to generate RSA key pair: %s", err)
 		}
-		log.Println("Key generated")
-		log.Println("Extracting public key")
-		pk, err := rsaPub(ctx, session, pub, *rsaModLen, *rsaExp)
-		if err != nil {
-			log.Fatal(err.Error())
-		}
-		log.Println("Extracted public key")
-		log.Println("Verifying public key")
-		err = rsaVerify(ctx, session, priv, pk)
-		if err != nil {
-			log.Fatal(err.Error())
-		}
-		log.Println("Key verified")
-		pubKey = pk
 	case "ECDSA":
 		if *ecdsaCurve == "" {
 			panic("--ecdsaCurve is required")
 		}
-		curve, present := stringToCurve[*ecdsaCurve]
-		if !present {
-			panic("curve not supported")
-		}
-		log.Printf("Generating ECDSA key with curve %s\n", *ecdsaCurve)
-		m, pubTmpl, privTmpl := ecArgs(*label, curve)
-		pub, priv, err := ctx.GenerateKeyPair(session, m, pubTmpl, privTmpl)
+		pubKey, err = ecdsaGenerate(ctx, session, *label, *ecdsaCurve)
 		if err != nil {
-			log.Fatal(err)
+			log.Fatalf("Failed to generate ECDSA key pair: %s", err)
 		}
-		log.Println("Key generated")
-		log.Println("Extracting public key")
-		pk, err := ecPub(ctx, session, pub, curve)
-		if err != nil {
-			log.Fatal(err)
-		}
-		log.Println("Extracted public key")
-		log.Println("Verifying public key")
-		err = ecVerify(ctx, session, priv, pk)
-		if err != nil {
-			log.Fatal(err)
-		}
-		log.Println("Key verified")
-		pubKey = pk
 	}
 
 	der, err := x509.MarshalPKIXPublicKey(pubKey)
